Initial python package and tools setup
This commit sets up this repository as a python package using `uv` for
dependency management, `ruff` for formatting and linting, and `mypy` for
type checking. To make it easier to remember how to invoke these tools
in the correct environment, I added wrapper shell scripts in the dev
directory that call them correctly.

Tested by:
- `./dev/autofix_lint.sh`
- `./dev/format.sh`
- `./dev/lint.sh`
- `./dev/type_check.sh`