Set up the computed stitching parameters
There's a bunch of parameters used in the stitching code that are
neither part of the image processing code per se, nor required as user
inputs. Formerly, these were just set up as extra properties on the
Stitcher class in a bunch of different functions (I started from:
https://github.com/sohamazing/image-stitcher/blob/622b06e5a9151c35453ea03ffc9e1b607547159c/stitcher.py)

To make it clearer what these all are supposed to be, test their
computation, and clean up the large Sticher class, I've moved them out
into their own parameters dataclass. To safeguard against using the
default values that were there before but never intended to be actually
used, I lightly restructured it so that the fields use
`field(init=False)` for the default, which will raise an attribute if a
later bug means that we didn't initialize a field that we expected to be
initialized. (It might be worth doing additional refactoring later to
avoid this existing late initalization pattern altogether.)

This also includes a change to change the dask array chunk size to
ensure that it's at least the size of a single image plane. In my
testing, this was a very large speedup (up to 25x for the largest image
I tested, which was 400 fov) for image planes that can fit in memory and
thus were being chunked needlessly.

Finally, I added some basic tests for computing some of the parameters
to make it easier to make changes confidently in the future.

Tested by:
- `./dev/autofix_lint.sh`
- `./dev/format.sh`
- `./dev/type_check.sh`
- `uv run python -m unittest image_stitcher.parameters_test`