Use skimage for image reading direct to numpy arrays instead of dask
When reading the raw images from disk, we're reading single fields of
view for a single region and timepoint, so we should not have to worry
about images not fitting into memory, and we can cut out dask from this
step. This both makes profiling easier and also in my testing cuts off
about 1s from the 8x8 field of view benchmark I've been using (which is
down down in the 10-11s range).

Tested by:
- `./dev/run_tests.sh`